fix/#184/이벤트 상수 모으기
